requirements.txt
Flask
Flask-Cors
peewee
passlib
PyJWT
psycopg2-binary
python-dotenv
